AshArchival.Resource.Info (ash_archival v1.0.3)

Introspection helpers for AshArchival.Resource

Functions

archive_archive_related(dsl_or_extended)

@spec archive_archive_related(dsl_or_extended :: module() | map()) ::
  {:ok, [atom()]} | :error

A list of relationships that should have all related items archived when this is archived. Notifications are not sent for this operation.

archive_archive_related!(dsl_or_extended)

@spec archive_archive_related!(dsl_or_extended :: module() | map()) ::
  [atom()] | no_return()

A list of relationships that should have all related items archived when this is archived. Notifications are not sent for this operation.

archive_attribute(dsl_or_extended)

@spec archive_attribute(dsl_or_extended :: module() | map()) :: {:ok, atom()} | :error

The attribute in which to store the archival flag (the current datetime).

archive_attribute!(dsl_or_extended)

@spec archive_attribute!(dsl_or_extended :: module() | map()) :: atom() | no_return()

The attribute in which to store the archival flag (the current datetime).

archive_base_filter?(dsl_or_extended)

@spec archive_base_filter?(dsl_or_extended :: module() | map()) :: atom()

Whether or not a base filter exists that applies the is_nil(archived_at) rule.

archive_exclude_destroy_actions(dsl_or_extended)

@spec archive_exclude_destroy_actions(dsl_or_extended :: module() | map()) ::
  {:ok, [atom()]} | :error

A destroy action or actions that should not archive, but instead be left alone. This allows for having a destroy or archive pattern.

archive_exclude_destroy_actions!(dsl_or_extended)

@spec archive_exclude_destroy_actions!(dsl_or_extended :: module() | map()) ::
  [atom()] | no_return()

A destroy action or actions that should not archive, but instead be left alone. This allows for having a destroy or archive pattern.

archive_exclude_read_actions(dsl_or_extended)

@spec archive_exclude_read_actions(dsl_or_extended :: module() | map()) ::
  {:ok, [atom()]} | :error

A read action or actions that should show archived items. They will not get the automatic is_nil(archived_at) filter.

archive_exclude_read_actions!(dsl_or_extended)

@spec archive_exclude_read_actions!(dsl_or_extended :: module() | map()) ::
  [atom()] | no_return()

A read action or actions that should show archived items. They will not get the automatic is_nil(archived_at) filter.

archive_exclude_upsert_actions(dsl_or_extended)

@spec archive_exclude_upsert_actions(dsl_or_extended :: module() | map()) ::
  {:ok, [atom()]} | :error

This option is deprecated as it no longer has any effect. Upserts are handled according to the upsert identity. See the upserts guide for more.

archive_exclude_upsert_actions!(dsl_or_extended)

@spec archive_exclude_upsert_actions!(dsl_or_extended :: module() | map()) ::
  [atom()] | no_return()

This option is deprecated as it no longer has any effect. Upserts are handled according to the upsert identity. See the upserts guide for more.

archive_options(dsl_or_extended)

@spec archive_options(dsl_or_extended :: module() | map()) :: %{
  required(atom()) => any()
}

archive DSL options

Returns a map containing the and any configured or default values.